third_party/WebKit/LayoutTests/fast/js/string-match-expected.txt

String.match(…) test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testMatch(/x/g, false) is ["x","x"]
FAIL testMatch(/x/g, true) should throw an exception. Was x,x.
PASS testMatch(/x/, false) is ["x"]
PASS testMatch(/x/, true) is ["x"]
PASS testMatch(/x/g, false); re.lastIndex is 0
FAIL testMatch(/x/g, true); re.lastIndex should throw an exception. Was 3.
PASS testMatch(/x/, false); re.lastIndex is 3
PASS testMatch(/x/, true); re.lastIndex is 3
PASS successfullyParsed is true

TEST COMPLETE
